move FamilyListTable folder from pages folder into common folder
